audio: reduce stuttering on scene change in webgl
The WebGL implementation of FMOD is finicky, mostly due to some of the
core implementation details of HTML5. The main consequence for this
project is that audio notably stuttered on scene changes, but only in
the WebGL builds. This commit tries to alleviate that by optimizing the
size of streamed audio and tweaking the DSP buffers.

Signed-off-by: Jose A. Rivera <jarrpa@gmail.com>